% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{db_read}
\alias{db_read}
\title{Read database tables}
\usage{
db_read(src = NULL, tables = NULL, close = !is.dbi(src), ...)
}
\arguments{
\item{src}{A \code{DBIConnection} (dbi) object, as returned by \code{\link[=db_open]{db_open()}} or
data.base \code{path}
\itemize{
\item \code{dbi}: not close after reading (default)
\item \code{path}: close after reading (default)
}}

\item{tables}{vector of dbTable names. If not specified, read the first table.}

\item{close}{Boolean. Whether close connection after read.
\itemize{
\item \code{opened}: not close
\item \verb{not opened}: close
}}

\item{...}{Other parameters passed on to methods.}
}
\description{
Read database tables
}
