% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{year}
\alias{month}
\alias{day}
\alias{hour}
\alias{fread}
\alias{fwrite}
\alias{data.table}
\alias{str_extract}
\alias{tbl}
\alias{is.tbl}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{data.table}}, \code{\link[data.table]{fread}}, \code{\link[data.table]{fwrite}}}

  \item{dplyr}{\code{\link[dplyr:tbl]{is.tbl}}, \code{\link[dplyr]{tbl}}}

  \item{lubridate}{\code{\link[lubridate]{day}}, \code{\link[lubridate]{hour}}, \code{\link[lubridate]{month}}, \code{\link[lubridate]{year}}}

  \item{stringr}{\code{\link[stringr]{str_extract}}}
}}

