% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_curlR.R
\name{write_db}
\alias{write_db}
\alias{write_db_char}
\alias{write_db_batch}
\title{write_db}
\usage{
write_db(
  con,
  d,
  time,
  table_data = "runoff_hourly",
  table_info = "timeinfo",
  log = TRUE,
  check_time = TRUE,
  overwrite = FALSE,
  append = TRUE,
  mink = 10
)

write_db_char(con, file, overwrite = FALSE, append = TRUE, mink = 10)

write_db_batch(dbname = "chinawater.db", files, overwrite = TRUE)
}
\arguments{
\item{con}{SQLite connection returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}, or the file path of SQLite.}

\item{d}{data.frame or data.table object.}

\item{time}{POSIXct or time character, e.g. "1998-01-01 01:00:00"}

\item{mink}{minimum records}
}
\description{
write_db
}
\examples{
# write_db_batch("chinawater.db", files)
}
