% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_open.R
\name{db_open}
\alias{db_open}
\alias{db_close}
\title{open data.base}
\usage{
db_open(con = ":memory:", ...)

db_close(con = NULL)
}
\arguments{
\item{...}{authentication arguments needed by the DBMS instance; these
typically include \code{user}, \code{password}, \code{host}, \code{port}, \code{dbname}, etc.
For details see the appropriate \code{DBIDriver}.}
}
\description{
open data.base
}
\examples{
library(dplyr)
con <- db_open()
DBI::dbWriteTable(con, "mtcars", mtcars)
tbl(con, "mtcars")

}
\seealso{
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}
}
