% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{db_read}
\alias{db_read}
\title{Read database tables}
\usage{
db_read(con = NULL, tables = NULL, close = TRUE, ...)
}
\arguments{
\item{con}{A \code{DBIConnection} object, as returned by \code{\link[=db_open]{db_open()}} or
data.base path}

\item{tables}{vector of dbTable names. If not specified, read the first table.}

\item{...}{Other parameters passed on to methods.}
}
\description{
Read database tables
}
