% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_read.R
\name{db_read}
\alias{db_read}
\title{Read database tables}
\usage{
db_read(conn, tables, ...)
}
\arguments{
\item{conn}{A \code{DBIConnection} object, as returned by \code{\link[=db_open]{db_open()}} or
data.base path}

\item{...}{Other parameters passed on to methods.}
}
\description{
Read database tables
}
